More Propagation Debugging
The issue seems to be that when a troop is added through the propagatiojn order, it isn't added properly and thus does not function as a troop. Using casting I seem to have got it to propagate but now, during a conquest order, it tries to integrate in the next time period and fails on integratinging that troop. It seems that whatever object it is odes not have a DoIntegration function. Maybe doing a deep copy when creating the propagation order could work. Will have to look at everywhere propagation orders are created and try making things deep copies.