Dark Ages
Added functionality for a new modifier for the power values of the time periods which picks a point somewhere in the middle of the timeline for each planet and scales the power of the time periods based on how far they are from that randomly determined low point: dark age. This can be enabled or disabled with a new tunable value called darkAges.