Troop Propagation Health Fix
Fixed an issue where troops which propagated would appear in the next time period with full health even if they had taken damage in the original time period. This involved adding a new parameter to the troop constructor and checking in the constructor if that parameter was specified and, if so, using the supplied health, and if not using the default full health.