Multiplayer Work Continued
Added a function to server.ts to initialize starting game values when the server starts. Also added a get that returns the gamestate. It works by filling in a custom json file and sending it. Added functionality to the client to fetch the gamestate from the server and fill in the values to make the game work. Run the fetch at client startup so every time the player refreshes they get the gamestate fresh. Added a bash script to launch the server. Deleted an unnecessary data file from the server.