Time Period Armies Gamestate
The Armies in each time period are now sent in the gamestate to the client and the client loads them in. The default aarmy created in the time period constructor has been removed from the client side as it was creating an aarmy when making the time period to fill in from the gamestate json.