Turn Submission Continued
Added the refresh gamestate button. Added functionality to the client to send over the lists of troops involved in trades. Made it so buttons are hidden if it is not your turn. Added functionality to drawboard to hide certain things if the player is a spectator. Moved the gamestate fetch code to its own function so it can be reused by the refresh button. Added advance turn function to the server. Added a trade function to the server, currently empty, will need to be filled out. Set the server to run startturn for the first player in initialize. Added server functionality for trades coming in from player. Submitting a turn crashes the server, bug to be fixed.