Refresh bug fix
Fixed a bug with refreshing the game state where the gamestate on the client would not be fully reset and it was adding the list of players to the player list a second time instead of clearing the player list first. The player list and planet list are now cleared on refresh and filled when the gamestate is fetched.