Combat bug fix
Fixed a bug with trading on the server side where the new army created for a player moving troops into a time period would be created with the wrong owner index which would cause the client to crash when trying to read which player owns the time period. Submitting a turn where you train troops causes the client to crash now.