Fixes and UI
Added an element to the time period boxes to show if there is an active war in the time period. For now it is just the word war like the owner name. Fixed a bug where removing troops from a time period would not remove them through propagation in future time periods. Made it so troop removal propagation will remove all copies of a troop in a time period instead of just one.