Time period Board UI
Improved the styling of the time period board to show the owner and if there is combat in a clearer way and in a way thatapproximated the eventual flip board look. Fixed a bug where starting a war in a time period that was just conquested can prevent the conquest from propagating.