Button Styling and Logging Improvements
Added some styling to make buttons look better. Changed the server launch script to create a folder with the date and time and put the logs in there so I can keep logs from previous tests and know when they are from. Changed the git ignore to reflect this change in log location.