Attrition Logic
The logic for recovery has been changed to now consume resources for each troop equal to that troops modifer. If the required resources are not available the troop instead takes the healthRecoveryPercent as damage instead of recovery as attrition. This should hopefully mean that death stacks stick around less as they need tons of resources and if they dont get them the troops will die in a few turns, 5 by default.